오버라이딩

+++ src/com/khyuna0/ch07/part01/Shape.java
@@ -6,8 +6,8 @@ public class Shape { // 부모 클래스
 	int height;
 	String color;
 	
-	public int area() { // 도형의 면적 계산 메소드
-		int area = width * height;
+	public double area() { // 도형의 면적 계산 메소드
+		double area = width * height;
 		return area; 
 	}
 }
